Add timestamps for updates. Prevent double adding hotspots.

+++ app.py
@@ -4,6 +4,7 @@ import pandas
 import os
 from datetime import timedelta
 import datetime
+import json
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', 'dev')
@@ -29,16 +30,26 @@ def exists_in_db(name):
 def update_status(to_disable, to_enable):
     conn = get_db_connection()
     for disable in to_disable:
-        conn.execute('INSERT INTO hotspots (call_number, is_disabled) VALUES (?, ?)', (disable, True))
-        conn.commit()
+        # prevent adding the same hotspot
+        hotspot = conn.execute('SELECT * FROM hotspots WHERE call_number = ?', (disable,)).fetchone()
+        if hotspot is None:
+            conn.execute('INSERT INTO hotspots (call_number, is_disabled) VALUES (?, ?)', (disable, True))
+            conn.commit()
     for enable in to_enable:
         conn.execute('UPDATE hotspots SET is_disabled = ?' 'WHERE call_number = ?', (False, enable['call_number']))
         conn.commit()
+
+    set_timestamp()
     conn.close()
 
 def delete_enabled():
     conn = get_db_connection()
-    conn.execute('DELETE FROM hotspots WHERE is_disabled = ?', (False, ))
+    # if there were enabled hotspots
+    enabled_count = conn.execute('SELECT COUNT(*) FROM hotspots WHERE is_disabled = ?', (False, )).fetchone()[0]
+    if enabled_count > 0:
+        set_timestamp()
+        conn.execute('DELETE FROM hotspots WHERE is_disabled = ?', (False, ))
+    
     conn.commit()
     conn.close()
     
@@ -98,13 +109,27 @@ def search(query):
 
 to_disable = []
 to_enable = []
-last_updated = ""
+
+def get_last_updated():
+    with open('./data/time.json', 'r') as file:
+        data = json.load(file)
+        
+        return data["last_updated"]
+
+def set_timestamp():
+    # set last updated
+    data = {
+        "last_updated": datetime.datetime.now().strftime("%m/%d/%y %H:%M:%S %p"),
+    }
+
+    with open('./data/time.json', 'w') as file:
+        json.dump(data, file, indent=4) 
 
 @app.route('/', methods=['GET', 'POST'])
 def index():
-    if request.method == 'POST':
-        update_status(to_disable, to_enable)
-    
+    last_updated = get_last_updated()
+    print(last_updated)
+
     results = []
 
     if request.args.get('is_searching'):
@@ -115,8 +140,14 @@ def index():
         results = conn.execute('SELECT * FROM hotspots').fetchall()
         conn.close()
     
-    return render_template('index.html', hotspots = results)
+    return render_template('index.html', hotspots = results, last_updated = last_updated)
     
+@app.route('/update', methods=['POST'])
+def update():
+    if request.method == 'POST':
+        update_status(to_disable, to_enable)
+        flash("Applied Changes")
+    return redirect(url_for('index'))
 
 @app.route('/upload', methods=['GET', "POST"])
 def upload():
@@ -134,8 +165,10 @@ def upload():
 
 @app.route('/confirm_delete', methods=['GET', 'POST'])
 def confirm_delete():
+    
     if request.method == "POST":
         delete_enabled()
+        
         flash('Deleted all enabled hotspots!', 'success')
     
     return redirect(url_for('index'))
@@ -148,6 +181,7 @@ def toggle(id):
     conn.execute('UPDATE hotspots SET is_disabled = ?' 'WHERE id = ?', (new_val, id))
     conn.commit()
     conn.close()
+    set_timestamp()
     flash('"{}" status changed'.format(hotspot['call_number']), "success")
     return redirect(url_for('index'))
 
@@ -158,6 +192,7 @@ def delete(id):
     conn.execute('DELETE FROM hotspots WHERE id = ?', (id,))
     conn.commit()
     conn.close()
+    set_timestamp()
     flash('"{}" deleted'.format(hotspot['call_number']), 'success')
     return redirect(url_for('index'))
 
@@ -170,6 +205,7 @@ def create():
         conn.execute('INSERT INTO hotspots (call_number, is_disabled) VALUES (?, ?)', (call_number, is_disabled))
         conn.commit()
         conn.close()
+        set_timestamp()
         return redirect(url_for('index'))
     
 @app.route('/<int:id>/edit', methods=("POST", "GET"))
@@ -184,6 +220,7 @@ def edit(id):
         conn.execute('UPDATE hotspots SET call_number = ?, is_disabled = ?' 'WHERE id = ?', (call_number, is_disabled,id))
         conn.commit()
         conn.close()
+        set_timestamp()
         flash('"{}" updated'.format(hotspot['call_number']), 'success')
         return redirect(url_for('index'))
     return render_template('edit.html', hotspot=hotspot, is_disabled=is_disabled)
